+++ Application/Shared/DTOs/AcademicPeriod/AcademicPeriodDto.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Application.Shared.DTOs.AcademicPeriod
 {
-    public class AcademicPeriodDto : BaseDto<long>
+    public class AcademicPeriodDto : BaseDto<int>
     {
         public string Code { get; set; } = string.Empty;
         public DateTime StartDate { get; set; }
