+++ Domain/Entities/Modality.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Domain.Entities
 {
-    public class Modality : BaseEntity<long>
+    public class Modality : BaseEntity<int>
     {
         public string Code { get; set; } = string.Empty;
         public string? Description { get; set; }
@@ -15,6 +15,5 @@ namespace Domain.Entities
         public bool RequiresResearchHotbed { get; set; } = false;
         public bool RequiresRegisterModalityApproval { get; set; } = true;
         public new int? IdUserCreatedAt { get; set; }
-
     }
 }
